Keep fixing TestClient bug. Time to fix read logic for maxPlayerCount.

+++ GBClient/include/GBClient.h
@@ -25,7 +25,7 @@ public: ~TestClient();
 public: void Start();
 
 public: template<typename STRING> void InscribeStringToBuffer(const STRING& str, GBBuffer* pGbBuffer);
-private: void ClientConsumeGatheredMessage(GBBuffer& buffer, uint32_t receivedMessageSize);
+private: void ClientConsumeGatheredMessage(GBBuffer& buffer, uint8_t receivedMessageSize);
 private: void ConsumeGatheredMessage(ExpandedOverlapped* pEol, char* message, const uint32_t messageSize, int& readOffSet) override {}
 private: void* GetSessionInformationArray() override { return nullptr; }
 private: void GatheringMessage(char* message, uint8_t leftBytesToRecieve);
